Move getdatafromurl() to webextras, and cleanup the chefkoch code
