+++ gourmet/importers/webextras.py
@@ -1,4 +1,5 @@
 import socket, gourmet.threadManager, urllib2
+from gourmet.gdebug import warn
 from gettext import gettext as _
 DEFAULT_SOCKET_TIMEOUT=45.0
 URLOPEN_SOCKET_TIMEOUT=15.0
@@ -84,4 +85,30 @@ def get_url (url, suspendableThread):
     else:
         sock = url
         return read_socket_w_progress(sock,suspendableThread,_('Retrieving file'))
+
+
+def getdatafromurl(url, content_type_check=None):
+    """Download data from URL.
+    @param url: the URL to download
+    @ptype url: string
+    @param content_type_check: if non-empty, only return data if the
+      Content-Type header starts with the given string
+    @ptype content_type_check: string
+    @return: URL data or None, and the url (which may be redirected
+      to a new URL)
+    @rtype: tuple (string, string) or (None, string)
+    """
+    data = None
+    try:
+        sock = urllib2.urlopen(url)
+        url = sock.geturl()
+        if content_type_check:
+            content_type = sock.info().get('content-type', 'application/octet-stream')
+            if content_type.lower().startswith(content_type_check):
+                data = sock.read()
+        else:
+            data = sock.read()
+    except urllib2.URLError as msg:
+        warn("could not get data from URL %r: %s" % (url, msg))
+    return data, url
